Feat: #123 RT 만료 시 후처리 구현
