+++ src/components/common/Header.tsx
@@ -1,13 +1,29 @@
-import { Link, NavLink } from 'react-router-dom';
+import { Link, NavLink, useNavigate } from 'react-router-dom';
 import logo from '@assets/logo.svg';
 import { FaUserCircle } from 'react-icons/fa';
 import { FiHome } from 'react-icons/fi';
 import { useStore } from '@stores/useStore';
-import useLogout from '@hooks/useLogout';
+import { logout } from '@services/authService';
+import useToast from '@hooks/useToast';
 
 export default function Header() {
-  const { userInfo: userInfoData } = useStore();
-  const { handleLogout } = useLogout();
+  const { userInfo: userInfoData, onLogout, clearUserInfo } = useStore();
+  const navigate = useNavigate();
+  const { toastSuccess } = useToast();
+
+  const handleLogout = async () => {
+    try {
+      await logout();
+      onLogout();
+      clearUserInfo();
+      navigate('/signin', { replace: true });
+      setTimeout(() => {
+        toastSuccess('로그아웃이 완료되었습니다.');
+      }, 100);
+    } catch (error) {
+      console.error('로그아웃 도중 에러가 발생했습니다.');
+    }
+  };
 
   return (
     <header className="flex h-header items-center justify-between bg-main px-15">
